Provide explicit lifetime for transmute results
Prior to this change, the calls to std::mem::transmute produced an
unbounded lifetime.

This change explicitly provides the lifetime parameter to the Chars
reference.  In the case of a new lexer this lifetime is bound to the
lexer itself.  Hence a lifetime parameter is also introduced, which in
turn bound to the two chars iterators contained within it.

Also, the type of the input parameter to Lexer::new is changed from a
String to a borrowed string slice (`&str`); this allows users to take
advantage of Rust's deref coercion and is more idiomatic of Rust.

Signed-off-by: Corin Lawson <corin@responsight.com>