Avoid calls to transmute
Prior to this change, std::mem::transmute was used to circumvent Rust's
borrow-checker.  This function is *incredibly* unsafe... best to avoid
it.  The call was doing two things: (1) it allowed the input string to
be borrowed after being consumed by the call to count; and (2) created a
new unbounded lifetime for the returned char iterator.

This change introduces a lifetime parameter to the lexer struct which
can then be bound to the lifetime of the char iterator.  Also, we clone
the iterator before calling count.  These measures overcome the two
problems above.

https://doc.rust-lang.org/std/mem/fn.transmute.html
https://doc.rust-lang.org/nomicon/unbounded-lifetimes.html
Signed-off-by: Corin Lawson <corin@responsight.com>